+++ console/src/routes.js
@@ -34,6 +34,7 @@ import ApiContainer from './components/Services/ApiExplorer/Container';
 import metadataOptionsConnector from './components/Services/Settings/MetadataOptions/MetadataOptions';
 import metadataStatusConnector from './components/Services/Settings/MetadataStatus/MetadataStatus';
 import allowedQueriesConnector from './components/Services/Settings/AllowedQueries/AllowedQueries';
+import inheritedRolesConnector from './components/Services/Settings/InheritedRoles/InheritedRoles';
 import logoutConnector from './components/Services/Settings/Logout/Logout';
 import aboutConnector from './components/Services/Settings/About/About';
 
@@ -157,6 +158,7 @@ const routes = store => {
           />
           <Route path="logout" component={logoutConnector(connect)} />
           <Route path="about" component={aboutConnector(connect)} />
+          <Route path="inherited-roles" component={inheritedRolesConnector} />
         </Route>
         {dataRouter}
         {remoteSchemaRouter}