+++ CHANGELOG.md
@@ -10,6 +10,8 @@
 - server: inherited roles for PG queries and subscription
 - server: fix issue when a remote relationship's joining field had a custom GraphQL name defined (fix #6626)
 - server: fix handling of nullable object relationships (fix #6633)
+- console: add inherited roles support (#483)
+- console: add permissions support for mssql tables (#677)
 - cli: support rest endpoints
 - cli: support mssql sources
 - cli: use relative paths in metadata !include directives
@@ -27,7 +29,6 @@
 - server/mssql: fix text values erroneously being parsed as varchar
 - server: improve errors messages for inconsistent sources
 - console: add relationship tab for mssql tables (#677)
-- console: add permissions support for mssql tables (#677)
 - build: fix the packaging of static console assets (fix #6610)
 - server: make REST endpoint errors compatible with inconsistent metadata